ANDA
Anda todo

+++ FinalSIM_Lavadero/Clases/Automovil.vb
@@ -1,3 +1,3 @@
 ﻿Public Class Automovil
-    Public num As Int32, estado As String, tipo As String, numK As Double, tiempoLlegada As Double
+    Public num As Int32, estado As String, tipo As String, numK As Single, tiempoLlegada As Single
 End Class
